[core]todoList : check task is_done , start login form

+++ tpl/tpl-index.php
@@ -12,7 +12,7 @@
 <div class="page">
   <div class="pageHeader">
     <div class="title">Dashboard</div>
-    <div class="userPanel"><i class="fa fa-chevron-down"></i><span class="username">John Doe </span><img src="https://s3.amazonaws.com/uifaces/faces/twitter/kolage/73.jpg" width="40" height="40"/></div>
+    <div class="userPanel"><i class="fa fa-chevron-down"></i><span class="username">John Doe </span><img src="#" width="40" height="40"/></div>
   </div>
   <div class="main">
     <div class="nav">
@@ -24,15 +24,18 @@
       <div class="menu">
         <div class="title">Folders</div>
         <ul class="folder-list">
-          <?php foreach ($folders as $folder): ?>
-          <li> 
-            <a href="?folder_id=<?= $folder->id ?>"><i class="fa fa-folder"></i><?= $folder->name ?></a>
-            <a href="?delete_folder=<?= $folder->id ?>" class="remove"> X </a>
-          </li>
-            <?php endforeach; ?>
+        <li class="<?= isset($_GET['folder_id']) ? '' : 'active' ?>">
+          <a href="<?= site_url()?>"><i class="fa fa-tasks"></i> ALL</a>
+        </li>
 
+          <!-- <?php foreach ($folders as $folder): ?> -->
+            <li class="<?= ($_GET["folder_id"] == $folder->id) ? 'active' : '' ?>">
+              <a href="?folder_id=<?= $folder->id ?>"><i class="fa fa-folder"></i><?= $folder->name ?></a>
+              <a href="?delete_folder=<?= $folder->id ?>" class="remove" onclick='return confirm("are you sure to delete ? ");'> X </a>
+            </li>
+            <!-- <?php endforeach; ?> -->
 
-          <li class="active"> <i class="fa fa-tasks"></i>Current Folder</li>
+ 
 
         </ul>
       </div>
@@ -43,38 +46,41 @@
     </div>
     <div class="view">
       <div class="viewHeader">
-        <div class="title">Manage Tasks</div>
+        <div class="title">
+        <input type="text" id="taskNameInput" placeholder="taskNameInput"/>
+      
+      </div>
         <div class="functions">
           <div class="button active">Add New Task</div>
           <div class="button">Completed</div>
-          <div class="button inverz"><i class="fa fa-trash-o"></i></div>
         </div>
       </div>
       <div class="content">
         <div class="list">
           <div class="title">Today</div>
           <ul>
-            <li class="checked"><i class="fa fa-check-square-o"></i><span>Update team page</span>
+
+        <!-- <?php if (sizeof($tasks)): ?> -->
+
+          <?php foreach ($tasks as $task): ?>
+            <li class="<?= $task->is_done ? 'checked' : '' ; ?>"> 
+              
+              <i data-taskId="<?= $task->id?>" class="isDone fa <?= $task->is_done ? 'fa-check-square-o' : 'fa-square-o' ; ?>"></i>
+            <!-- <?php var_dump($task);?> -->
+            <span><?= htmlspecialchars($task->title) ?></span>
+
               <div class="info">
-                <div class="button green">In progress</div><span>Complete by 25/04/2014</span>
+                <span style='color: cadetblue;   font-size: 13px;   margin-right: 12px;'><?= $task->created_at ?></span>
+                <a href="?delete_task=<?= $task->id ?>" class="remove" onclick='return confirm("are you sure to delete ? ");'> X </a>
               </div>
             </li>
-            <li><i class="fa fa-square-o"></i><span>Design a new logo</span>
-              <div class="info">
-                <div class="button">Pending</div><span>Complete by 10/04/2014</span>
-              </div>
+            <?php endforeach; ?>
+          <?php else: ?>
+            <li>no task here .. 
             </li>
-            <li><i class="fa fa-square-o"></i><span>Find a front end developer</span>
-              <div class="info"></div>
-            </li>
-          </ul>
-        </div>
-        <div class="list">
-          <div class="title">Tomorrow</div>
-          <ul>
-            <li><i class="fa fa-square-o"></i><span>Find front end developer</span>
-              <div class="info"></div>
-            </li>
+          <?php endif; ?>
+
+
           </ul>
         </div>
       </div>
@@ -85,7 +91,21 @@
 <script src='//cdnjs.cloudflare.com/ajax/libs/jquery/2.1.3/jquery.min.js'></script><script  src="assets/js/script.js"></script>
 
 <script>    
+
   $(document).ready(function(){
+
+
+    $('.isDone').click(function(e){
+      var tid= $(this).attr('data-taskId');
+      $.ajax({
+          url: "process/ajaxHandler.php",
+          method: "POST",
+          data: {action: "doneSwitch", taskId: tid},
+          success: function(response) {
+              location.reload();
+          }
+        });
+    });
       $('#addFolderBtn').click(function(e){
         var input = $('#addFolderInput');
         // alert(input.val());
@@ -103,7 +123,25 @@
             }
           }
         });
+
+        });
+        $('#taskNameInput').on('keypress',function(e) {
+          if(e.which == 13) {
+            $.ajax({
+            url: "process/ajaxHandler.php",
+            method: "POST",
+            data: {action: "addTask",folderId: <?= $_GET['folder_id'] ?? 0?> ,taskTitle: $('#taskNameInput').val()},
+            success: function(response) {
+              
+              if(response == '1'){
+                  location.reload();
+              }else{
+                alert(response);
+              }
+            }
+        });          }
     });
+    $('#taskNameInput').focus();
   });
     </script>
